requirements.txt
alembic==1.6.5
ariadne==0.13.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
graphql-core==3.1.5
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
numpy==1.21.1
pandas==1.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
